Update ECR deployment workflow to use consistent image naming conventions for backend and frontend
This commit modifies the ecr-deploy-stg GitHub Actions workflow to change the image names for the backend and frontend Docker images from 'api:stg' and 'front:stg' to 'stg-backend:latest' and 'stg-frontend:latest', respectively. This change enhances clarity and consistency in the image naming strategy used during the build and push processes.