Update ECR deployment workflow to use secrets for AWS region and registry
This commit modifies the ecr-deploy-stg GitHub Actions workflow to replace environment variables with secrets for the AWS region and ECR registry. This change enhances security by ensuring sensitive information is not exposed in the workflow, while maintaining the functionality of the build and deployment processes for both backend and frontend images.